moving case config to a global place
